pyproject.toml
[tool.poetry]
name = "infiscal-secrets"
version = "0.1.0"
description = "Secrets Manager"
authors = ["orionhunts-ai <orion@oriondefensiveai.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Quart = "^0.19.6"
infisical = "^1.6.0"
infisical-python = "^2.2.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
